fuse: add write op
